Jake's Edits
Changes made to Branding, CSS, Page Layout etc. PLEASE NOTE, THE "/img" folder is NOT Included in this upload as GitHub prevents me from uploading that many files at once. I DID NOT modify anything in the images folder so if you merge the existing /img folder into this it should work as before.